NFFT-2: Generalized floating-point data type in a few other places. Some changes to avoid compiler warnings.
git-svn-id: https://euler.mathematik.tu-chemnitz.de/repos/nfft/lib/branches/nfft-3.3@4248 c8b3e4ef-9df4-0310-9d76-b6dd82e1323a

+++ kernel/util/window.c
@@ -37,6 +37,6 @@
  */
 INT Y(m2K)(const INT m)
 {
-  int j = MIN(m, (sizeof(m2K_) / sizeof(m2K_[0])) - 1);
+  int j = MIN(((int)(m)), ((int)((sizeof(m2K_) / sizeof(m2K_[0])) - 1)));
   return (INT)((1U << m2K_[j]) * (m + 2));
 }
